arrange files and skip debounce on first load
